requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
pdfplumber==0.11.4
pypdf==4.2.0
openai==1.55.3